array linear search
In Linear Search, we look for an element or value in an array by traversing it from the beginning to the end until the required element or value is discovered. The array is searched progressively, and if the key element to be searched is found in the array, the position is returned; otherwise, -1 is returned.